requirements.txt
# avalon_backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
httpx>=0.26.0
beautifulsoup4>=4.12.3
playwright>=1.41.0
apscheduler>=3.10.4